Delete the entire project.
I'm going to have it be ObjC based with some Swift classes inside.
I would prefer to learn Swift inside of a safer area, rather than
doing the entire app in Swift. I want to be able to make some progress
on the project or it's going to be hard to move forward on this in
my free time.