Change dashboard to use a group with a background image. Move summary label. Set empty state.
If there are no checkpoints, the dashboard prompts the user to get started.
I moved the summary image to be a background image of a group so that I could
have a text label over the top of it. Hopefully the background image lets me
do animation. If it doesn't, I will need to explore other routes or maybe just
put it back the way it was.